+++ tests/blackboard.rs
@@ -1,8 +1,10 @@
+//! Tests for [`BlackboardData`] and [`SharedBlackboard`]/[`Blackboard`](behaviortree::Blackboard)
 // Copyright © 2025 Stephan Kunz
 
-//! Tests for [`Blackboard`] and [`BlackboardNode`]
+#![allow(missing_docs)]
+#![allow(clippy::unwrap_used)]
 
-use behaviortree::{BlackboardData, BlackboardInterface, SHOULD_NOT_HAPPEN, SharedBlackboard, port::PortRemappings};
+use behaviortree::{BlackboardData, BlackboardInterface, SharedBlackboard, port::PortRemappings};
 
 #[test]
 fn blackboard() {
@@ -13,40 +15,30 @@ fn blackboard() {
 
 	let old = blackboard
 		.set("test", String::from("test"))
-		.expect(SHOULD_NOT_HAPPEN);
+		.unwrap();
 	assert_eq!(old, None);
 
-	let value = blackboard
-		.get::<String>("test")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = blackboard.get::<String>("test").unwrap();
 	assert_eq!(value, String::from("test"));
 
 	let old = blackboard
 		.set("test", String::from("changed"))
-		.expect(SHOULD_NOT_HAPPEN);
+		.unwrap();
 	assert_eq!(old, Some(String::from("test")));
 
-	let value = blackboard
-		.get::<String>("test")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = blackboard.get::<String>("test").unwrap();
 	assert_eq!(value, String::from("changed"));
 
 	let value = blackboard.set("test", 42);
 	assert!(value.is_err());
 
-	let old = blackboard
-		.delete::<String>("test")
-		.expect(SHOULD_NOT_HAPPEN);
+	let old = blackboard.delete::<String>("test").unwrap();
 	assert_eq!(old, String::from("changed"));
 
-	let old = blackboard
-		.set("test", 42)
-		.expect(SHOULD_NOT_HAPPEN);
+	let old = blackboard.set("test", 42).unwrap();
 	assert_eq!(old, None);
 
-	let value = blackboard
-		.get::<i32>("test")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = blackboard.get::<i32>("test").unwrap();
 	assert_eq!(value, 42);
 }
 
@@ -57,30 +49,22 @@ fn blackboard_node_default() {
 	let value = level0.get::<i32>("test");
 	assert!(value.is_err());
 
-	let old = level0
-		.set("test", String::from("test"))
-		.expect(SHOULD_NOT_HAPPEN);
+	let old = level0.set("test", String::from("test")).unwrap();
 	assert_eq!(old, None);
 
-	let value = level0
-		.get::<String>("test")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = level0.get::<String>("test").unwrap();
 	assert_eq!(value, String::from("test"));
 
 	let old = level0.set("test", 42);
 	assert!(old.is_err());
 
-	let old = level0
-		.delete::<String>("test")
-		.expect(SHOULD_NOT_HAPPEN);
+	let old = level0.delete::<String>("test").unwrap();
 	assert_eq!(old, String::from("test"));
 
-	let old = level0.set("test", 42).expect(SHOULD_NOT_HAPPEN);
+	let old = level0.set("test", 42).unwrap();
 	assert_eq!(old, None);
 
-	let value = level0
-		.get::<i32>("test")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = level0.get::<i32>("test").unwrap();
 	assert_eq!(value, 42);
 }
 
@@ -90,44 +74,30 @@ fn blackboard_node_with_parent() {
 
 	let old = level0
 		.set("test1", String::from("test1"))
-		.expect(SHOULD_NOT_HAPPEN);
+		.unwrap();
 	assert_eq!(old, None);
 	let old = level0
 		.set("test2", String::from("test2"))
-		.expect(SHOULD_NOT_HAPPEN);
+		.unwrap();
 	assert_eq!(old, None);
 
 	let mut remappings = PortRemappings::default();
-	remappings
-		.add("test", "test1")
-		.expect(SHOULD_NOT_HAPPEN);
+	remappings.add("test", "test1").unwrap();
 	let mut node = SharedBlackboard::with_parent("level0", level0, remappings, true);
 
-	let old = node
-		.set("@other", String::from("other"))
-		.expect(SHOULD_NOT_HAPPEN);
+	let old = node.set("@other", String::from("other")).unwrap();
 	assert_eq!(old, None);
 
-	let old = node
-		.set("test3", String::from("test3"))
-		.expect(SHOULD_NOT_HAPPEN);
+	let old = node.set("test3", String::from("test3")).unwrap();
 	assert_eq!(old, None);
 
-	let value = node
-		.get::<String>("@other")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = node.get::<String>("@other").unwrap();
 	assert_eq!(value, String::from("other"));
-	let value = node
-		.get::<String>("test")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = node.get::<String>("test").unwrap();
 	assert_eq!(value, String::from("test1"));
-	let value = node
-		.get::<String>("test2")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = node.get::<String>("test2").unwrap();
 	assert_eq!(value, String::from("test2"));
-	let value = node
-		.get::<String>("test3")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = node.get::<String>("test3").unwrap();
 	assert_eq!(value, String::from("test3"));
 }
 
@@ -136,110 +106,82 @@ fn blackboard_node_hierarchy() {
 	let mut level0 = SharedBlackboard::new("level0");
 
 	let mut remappings1 = PortRemappings::default();
-	remappings1
-		.add("levelB", "levelA")
-		.expect(SHOULD_NOT_HAPPEN);
+	remappings1.add("levelB", "levelA").unwrap();
 	let mut level1 = SharedBlackboard::with_parent("level1", level0.clone(), remappings1, true);
 
 	let mut remappings2 = PortRemappings::default();
-	remappings2
-		.add("levelC", "levelB")
-		.expect(SHOULD_NOT_HAPPEN);
+	remappings2.add("levelC", "levelB").unwrap();
 	let mut level2 = SharedBlackboard::with_parent("level2", level1.clone(), remappings2, true);
 
 	let mut remappings3 = PortRemappings::default();
-	remappings3
-		.add("levelD", "levelC")
-		.expect(SHOULD_NOT_HAPPEN);
+	remappings3.add("levelD", "levelC").unwrap();
 	let mut level3 = SharedBlackboard::with_parent("level3", level2.clone(), remappings3, true);
 
 	let old = level0
 		.set("levelA", String::from("testA"))
-		.expect(SHOULD_NOT_HAPPEN);
+		.unwrap();
 	assert_eq!(old, None);
 
 	let old = level0
 		.set("level0", String::from("test0"))
-		.expect(SHOULD_NOT_HAPPEN);
+		.unwrap();
 	assert_eq!(old, None);
 	let old = level1
 		.set("level1", String::from("test1"))
-		.expect(SHOULD_NOT_HAPPEN);
+		.unwrap();
 	assert_eq!(old, None);
 	let old = level2
 		.set("level2", String::from("test2"))
-		.expect(SHOULD_NOT_HAPPEN);
+		.unwrap();
 	assert_eq!(old, None);
 	let old = level3
 		.set("level3", String::from("test3"))
-		.expect(SHOULD_NOT_HAPPEN);
+		.unwrap();
 	assert_eq!(old, None);
 
 	// test autoremap
 	let old = level3
 		.set("level2", String::from("changed2"))
-		.expect(SHOULD_NOT_HAPPEN);
+		.unwrap();
 	assert_eq!(old, Some(String::from("test2")));
-	let value = level3
-		.get::<String>("level2")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = level3.get::<String>("level2").unwrap();
 	assert_eq!(value, String::from("changed2"));
-	let value = level3
-		.get::<String>("level1")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = level3.get::<String>("level1").unwrap();
 	assert_eq!(value, String::from("test1"));
-	let value = level3
-		.get::<String>("level0")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = level3.get::<String>("level0").unwrap();
 	assert_eq!(value, String::from("test0"));
 	let old = level3
 		.set("level0", String::from("changed0"))
-		.expect(SHOULD_NOT_HAPPEN);
+		.unwrap();
 	assert_eq!(old, Some(String::from("test0")));
-	let value = level3
-		.get::<String>("level0")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = level3.get::<String>("level0").unwrap();
 	assert_eq!(value, String::from("changed0"));
 	let old = level3.set("level0", 42);
 	assert!(old.is_err());
-	let _ = level3
-		.delete::<String>("level0")
-		.expect(SHOULD_NOT_HAPPEN);
+	let _ = level3.delete::<String>("level0").unwrap();
 	assert_eq!(value, String::from("changed0"));
-	let old = level3.set("level0", 42).expect(SHOULD_NOT_HAPPEN);
+	let old = level3.set("level0", 42).unwrap();
 	assert_eq!(old, None);
-	let value = level1
-		.get::<i32>("level0")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = level1.get::<i32>("level0").unwrap();
 	assert_eq!(value, 42);
 
 	// test manual remapping
-	let value = level3
-		.get::<String>("levelD")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = level3.get::<String>("levelD").unwrap();
 	assert_eq!(value, String::from("testA"));
-	let value = level2
-		.get::<String>("levelC")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = level2.get::<String>("levelC").unwrap();
 	assert_eq!(value, String::from("testA"));
 	let old = level3
 		.set("levelD", String::from("changedD"))
-		.expect(SHOULD_NOT_HAPPEN);
+		.unwrap();
 	assert_eq!(old, Some(String::from("testA")));
-	let value = level1
-		.get::<String>("levelB")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = level1.get::<String>("levelB").unwrap();
 	assert_eq!(value, String::from("changedD"));
 	let old = level3.set("levelD", 42);
 	assert!(old.is_err());
-	let _ = level3
-		.delete::<String>("levelD")
-		.expect(SHOULD_NOT_HAPPEN);
+	let _ = level3.delete::<String>("levelD").unwrap();
 	assert_eq!(value, String::from("changedD"));
-	let old = level3.set("levelD", 42).expect(SHOULD_NOT_HAPPEN);
+	let old = level3.set("levelD", 42).unwrap();
 	assert_eq!(old, None);
-	let value = level1
-		.get::<i32>("levelB")
-		.expect(SHOULD_NOT_HAPPEN);
+	let value = level1.get::<i32>("levelB").unwrap();
 	assert_eq!(value, 42);
 }